Spring security has been integrated

+++ src/main/java/com/aman/projects/airBnbApp/service/HotelServiceImpl.java
@@ -5,13 +5,16 @@ import com.aman.projects.airBnbApp.dto.HotelInfoDto;
 import com.aman.projects.airBnbApp.dto.RoomDto;
 import com.aman.projects.airBnbApp.entity.Hotel;
 import com.aman.projects.airBnbApp.entity.Room;
+import com.aman.projects.airBnbApp.entity.User;
 import com.aman.projects.airBnbApp.exceptions.ResourceNotFoundException;
+import com.aman.projects.airBnbApp.exceptions.UnAuthorizedException;
 import com.aman.projects.airBnbApp.repository.HotelRepository;
 import com.aman.projects.airBnbApp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,6 +34,8 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating a new hotel with name: {}",hotelDto.getName());
         Hotel hotel=modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
         hotel=hotelRepository.save(hotel);
         log.info("Created a new Hotel with ID:{}",hotel.getId());
         return modelMapper.map(hotel,HotelDto.class);
@@ -42,6 +47,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel=hotelRepository.findById(id)
                 .orElseThrow
                         (()->new ResourceNotFoundException("There is no property Listed with the given id:"+id));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -51,6 +60,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel=hotelRepository.findById(id)
                 .orElseThrow
                         (()->new ResourceNotFoundException("There is no property Listed with the given id:"+id));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         modelMapper.map(hotelDto,hotel);
         hotel.setId(id);
         hotel=hotelRepository.save(hotel);
@@ -68,6 +81,10 @@ public class HotelServiceImpl implements HotelService{
                 .orElseThrow
                         (()->new ResourceNotFoundException("There is no property Listed with the given id:"+id));
         //TODO:to delete the future inventories for this hotel
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         for(Room room:hotel.getRooms()){
             inventoryService.deleteFutureInventories(room);
             roomRepository.deleteById(room.getId());
@@ -82,6 +99,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel=hotelRepository.findById(id)
                 .orElseThrow
                         (()->new ResourceNotFoundException("There is no property Listed with the given id:"+id));
+        User user=(User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         hotel.setActive(true);
         //TODO: Create Inventory for all the rooms of this hotel
         for(Room room:hotel.getRooms()){
